refactor dto

+++ src/main/java/com/miro/api/widgets/testtask/dto/WidgetResponseDTO.java
@@ -5,14 +5,14 @@ package com.miro.api.widgets.testtask.dto;
  */
 public class WidgetResponseDTO {
     private final String id;
-    private final Integer xCoordinate;
-    private final Integer yCoordinate;
-    private final Integer zIndex;
-    private final Integer height;
-    private final Integer width;
-    private final Integer updatedAt;
+    private final int xCoordinate;
+    private final int yCoordinate;
+    private final int zIndex;
+    private final int height;
+    private final int width;
+    private final long updatedAt;
 
-    public WidgetResponseDTO(String id, Integer xCoordinate, Integer yCoordinate, Integer zIndex, Integer height, Integer width, Integer updatedAt) {
+    public WidgetResponseDTO(String id, int xCoordinate, int yCoordinate, int zIndex, int height, int width, long updatedAt) {
         this.id = id;
         this.xCoordinate = xCoordinate;
         this.yCoordinate = yCoordinate;
@@ -26,27 +26,27 @@ public class WidgetResponseDTO {
         return id;
     }
 
-    public Integer getXCoordinate() {
+    public int getXCoordinate() {
         return xCoordinate;
     }
 
-    public Integer getYCoordinate() {
+    public int getYCoordinate() {
         return yCoordinate;
     }
 
-    public Integer getZIndex() {
+    public int getZIndex() {
         return zIndex;
     }
 
-    public Integer getHeight() {
+    public int getHeight() {
         return height;
     }
 
-    public Integer getWidth() {
+    public int getWidth() {
         return width;
     }
 
-    public Integer getUpdatedAt() {
+    public long getUpdatedAt() {
         return updatedAt;
     }
 }
